tests.
git-svn-id: svn://zon/projecten/zicht-libraries/ZichtFileManagerBundle/trunk@20989 0f35748b-8228-47f0-bbea-889afbbd9a42

+++ src/Zicht/Bundle/FileManagerBundle/FileManager/FileManager.php
@@ -83,8 +83,10 @@ class FileManager {
      */
     public function __destruct()
     {
-        foreach ($this->preparedPaths as $file) {
-            unlink($file);
+        if (!empty($this->preparedPaths)) {
+            foreach ($this->preparedPaths as $file) {
+                unlink($file);
+            }
         }
     }
